Refactored
Refactored front-end requests to DB to request to the server, rather than directly to the DB.
Refactored server with REST API to accept the new requests and forward them to a local server.
Added history display on the bottom of pages with a limit of 25 items.
Now the server can be run anywhere and the user simply needs to specify the ip and port for the db at runtime:
-p = port for server
-dbp = port for db
-dbip = ip address for db